requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Multi-Agent Orchestration
pyautogen==0.2.9
langchain==0.1.4
langchain-google-genai==0.0.6

# LLM APIs
google-generativeai==0.3.2
openai==1.10.0

# Async & HTTP
httpx==0.26.0
aiohttp==3.9.1

# Security & APIs
requests==2.31.0
python-dotenv==1.0.0

# Database (Session Storage)
redis==5.0.1

# Utils
python-json-logger==2.0.7
pydantic-core==2.14.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Week 4: Multi-modal processing (optional but recommended)
pillow==10.2.0
# pytesseract==0.3.10  # Requires Tesseract OCR installed on system

# Week 4: Monitoring (optional)
# prometheus-client==0.19.0
# sentry-sdk==1.39.0

